Comparison/requirements.txt
absl-py==0.7.1
appnope==0.1.0
astor==0.7.1
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
Brotli==1.0.7
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
Flask==1.0.2
future==0.17.1
gast==0.2.2
grpcio==1.20.1
gym==0.12.1
h5py==2.9.0
idna==2.8
imageio==2.5.0
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keras-rl==0.4.2
kiwisolver==1.0.1
lxml==4.3.3
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
mock==2.0.0
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
notebook==5.7.8
numpy==1.16.2
opencv-python==4.1.0.25
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pbr==5.2.0
pdfkit==0.6.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
ptyprocess==0.6.0
PyAutoGUI==0.9.44
pycparser==2.19
pyfiglet==0.8.post1
pygame==1.9.6
PyGetWindow==0.0.5
pyglet==1.3.2
Pygments==2.3.1
PyMsgBox==1.0.6
pyparsing==2.4.0
PyRect==0.1.4
pyrsistent==0.15.2
PyScreeze==0.1.21
python-dateutil==2.8.0
PyTweening==1.0.3
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.4
reportlab==3.5.17
requests==2.21.0
scikit-image==0.15.0
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
SoundFile==0.10.2
soupsieve==1.9
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==3.4.2
